+++ app/settings.py
@@ -1,7 +1,6 @@
 import logging
 from structlog import wrap_logger
 import os.path
-import pwd
 import requests
 from requests.packages.urllib3.util.retry import Retry
 from requests.adapters import HTTPAdapter
@@ -16,43 +15,39 @@ APP_ROOT = os.path.dirname(os.path.abspath(__file__))
 APP_TMP = os.path.join(APP_ROOT, 'tmp')
 
 # Default to true, cast to boolean
-SDX_STORE_URL = os.getenv("SDX_STORE_URL", "http://sdx-store:5000")
-SDX_SEQUENCE_URL = os.getenv("SDX_SEQUENCE_URL", "http://sdx-sequence:5000")
+SDX_STORE_URL = os.getenv("SDX_STORE_URL")
+SDX_SEQUENCE_URL = os.getenv("SDX_SEQUENCE_URL")
 
-FTP_HOST = os.getenv('CTP_FTP_HOST', 'pure-ftpd')
+FTP_HOST = os.getenv('CTP_FTP_HOST',)
 FTP_USER = os.getenv('CTP_FTP_USER')
 FTP_PASS = os.getenv('CTP_FTP_PASS')
 
-FTP_FOLDER = os.getenv('CTP_FTP_FOLDER', '/')
-FTP_HEARTBEAT_FOLDER = os.getenv('CTP_FTP_HEARTBEAT_FOLDER', '/heartbeat')
+FTP_FOLDER = os.getenv('CTP_FTP_FOLDER')
+FTP_HEARTBEAT_FOLDER = os.getenv('CTP_FTP_HEARTBEAT_FOLDER')
 
-SFTP_HOST = os.getenv("SFTP_HOST", "127.0.0.1")
-SFTP_PORT = os.getenv("SFTP_PORT", "22")
-SFTP_USER = os.getenv('SFTP_USER', pwd.getpwuid(os.getuid())[0])
-SFTP_PRIVATEKEY_FILENAME = os.getenv(
-    "SFTP_PRIVATEKEY_FILENAME", os.path.expanduser(os.path.join("~", ".ssh", "id_rsa"))
-)
-SFTP_PUBLICKEY_FILENAME = os.getenv(
-    "SFTP_PUBLICKEY_FILENAME", os.path.expanduser(os.path.join("~", ".ssh", "id_rsa.pub"))
-)
+SFTP_HOST = os.getenv("SFTP_HOST")
+SFTP_PORT = os.getenv("SFTP_PORT")
+SFTP_USER = os.getenv('SFTP_USER')
+SFTP_PRIVATEKEY_FILENAME = os.getenv("SFTP_PRIVATEKEY_FILENAME")
+SFTP_PUBLICKEY_FILENAME = os.getenv("SFTP_PUBLICKEY_FILENAME")
 
-RABBIT_QUEUE = os.getenv('CTP_NOTIFICATIONS_QUEUE', 'sdx-ctp-survey-notifications')
-RABBIT_EXCHANGE = os.getenv('RABBITMQ_EXCHANGE', 'message')
+RABBIT_QUEUE = os.getenv('CTP_NOTIFICATIONS_QUEUE')
+RABBIT_EXCHANGE = os.getenv('RABBITMQ_EXCHANGE')
 
 RABBIT_URL = 'amqp://{user}:{password}@{hostname}:{port}/{vhost}'.format(
-    hostname=os.getenv('RABBITMQ_HOST', 'rabbit'),
-    port=os.getenv('RABBITMQ_PORT', 5672),
-    user=os.getenv('RABBITMQ_DEFAULT_USER', 'rabbit'),
-    password=os.getenv('RABBITMQ_DEFAULT_PASS', 'rabbit'),
-    vhost=os.getenv('RABBITMQ_DEFAULT_VHOST', '%2f')
+    hostname=os.getenv('RABBITMQ_HOST'),
+    port=os.getenv('RABBITMQ_PORT'),
+    user=os.getenv('RABBITMQ_DEFAULT_USER'),
+    password=os.getenv('RABBITMQ_DEFAULT_PASS'),
+    vhost=os.getenv('RABBITMQ_DEFAULT_VHOST')
 )
 
 RABBIT_URL2 = 'amqp://{user}:{password}@{hostname}:{port}/{vhost}'.format(
-    hostname=os.getenv('RABBITMQ_HOST2', 'rabbit'),
-    port=os.getenv('RABBITMQ_PORT2', 5672),
-    user=os.getenv('RABBITMQ_DEFAULT_USER', 'rabbit'),
-    password=os.getenv('RABBITMQ_DEFAULT_PASS', 'rabbit'),
-    vhost=os.getenv('RABBITMQ_DEFAULT_VHOST', '%2f')
+    hostname=os.getenv('RABBITMQ_HOST2'),
+    port=os.getenv('RABBITMQ_PORT2'),
+    user=os.getenv('RABBITMQ_DEFAULT_USER'),
+    password=os.getenv('RABBITMQ_DEFAULT_PASS'),
+    vhost=os.getenv('RABBITMQ_DEFAULT_VHOST')
 )
 
 RABBIT_URLS = [RABBIT_URL, RABBIT_URL2]